Unify admin embed command
